seperating the logger middle ware to seperate file
