+++ 02-express-tutorial/app.js
@@ -1,15 +1,8 @@
 const express = require("express");
 const app = express();
+const logger = require("./logger.js");
 
 // req => middleware => res
-
-const logger = (req, res, next) => {
-  const method = req.method;
-  const url = req.url;
-  const time = new Date().getFullYear();
-  console.log(method, url, time);
-  next();
-};
 
 app.get("/", logger, (req, res) => {
   res.send("home page");
